Add Update 10 trade changes + Format for PEP8 (#6)
Update 10 Trade Changes
----------------------------

Frontier Developments released Update 10 for Odyssey and Horizons on 25 Jan 2022. The [release notes](https://forums.frontier.co.uk/threads/update-10-odyssey-and-horizons.597956/) included:

> - Faction consequences from selling commodities are now based on satisfying available demand at the Commodity Market
>   - You can now support a faction by buying commodities from it, based upon satisfying available supply at its Commodity Market
>   - When buying and selling, it's generally more positive to a faction to satisfy a higher supply and demand (green 3-bar icon) than a lower (red 1-bar icon), and it's generally more positive to trade more expensive commodities than cheaper ones

As with all BGS changes in **Elite Dangerous**, the exact changes are ambiguous and unknown. Therefore, the code changes present more information about the trade, including the average value of the "supply and demand". Presumably, this is the supply and demand "bracket", a number from 0 to 3 represented in the in-game user interface as bars, as implied by the last bullet point. I assume this is not the supply or demand value, a count of T of the commodity.

The old way of reporting market sell events was `X T trade at Y CR average profit per T` where **X** was the total T traded and **Y** average profile or loss per T. The new way is `A market sell(s). Total: B T and C CR profit. Average: D CR/T profit at bracket E` where **A** is the count of market sell events, B is the total T sold, C the total CR profile or less, D the average profit or loss per T and E is the average demand bracket. Market buy events follow a similar format and were not previously reported because they were not BGS-impacting. This also satisfies #1.

The code treats any buy and a profitable sale as supporting or PRO the owning minor faction. The code still treats an unprofitable trade as undermining or ANTI the owning minor faction. Further testing will help refine this and I expect further changes to both refine the output and improve the PRO/ANTI accuracy.

PEP8 Code Reformat
-----------------------

I reformatted the code to be consistent with PEP8. I have saved the lining rules used for flake8 in `.flake8`. This ensures consistent formatting today and in the future.

Minor Changes
-----------------

Changed the help text in the plug-in's settings tab to mention the 'Available Minor Factions' list box by name. This makes it explicit and, therefore, less confusing to some.